Changed the shuttle tile thruster updating to its own event, UpdateSpacedEvent
