+++ Content.Server/Shuttles/Events/UpdateSpacedEvent.cs
@@ -1,0 +1,11 @@
+using Content.Server.Shuttles.Components;
+
+namespace Content.Server.Shuttles.Events;
+
+/// <summary>
+/// Raised whenever a tile changes from spaced to unspaced or vice versa.
+/// </summary>
+public sealed class UpdateSpacedEvent : EntityEventArgs
+{
+
+}
